requirements.txt
Flask==2.0.1
Flask-WTF==0.15.1
python-nmap==0.6.1
